refactor: 숫자에 해당하는 역할, 일급 컬렉션(Numbers.class)에 할당하기
